+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 v2.2.3
 ------
+* Fix a ConcurrentModificationException when comparing against a dynamic constant.
 * Detect simple cases in which the super or this constructor isn't invoked properly.
 
 v2.2.2 (2022-05-01)